+++ termcs/__init__.py
@@ -1,0 +1,5 @@
+from . import termcs
+
+
+def run():
+    termcs.run()