+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.0] - 2023-03-29
+
+### Changed
+
+- Overall increased performance due to the usage of a proper table widget from Textual
+
+### Added
+
+- Added the ability to sort the table by some column
+
+### Removed
+
+- Remove `stop` keybinding
+
 ## [0.0.2] - 2023-01-13
 
 ### Changed
@@ -15,5 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
+[0.1.0]: https://github.com/Riyum/termcs/compare/v0.0.2...v0.1.0
 [0.0.2]: https://github.com/Riyum/termcs/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/Riyum/termcs/releases/tag/v0.0.1
+